refactor: move journal_images tests to sibling tests.rs (#1084)
